requirements.txt
langchain_community
streamlit
pathlib
langchain_huggingface
langchain_groq
langchain_chroma
langchain_cohere
pandas
watchdog